refactor(web): finish splitting task detail route (#82)
Pull the right-rail sidebar, the inline RunningAgentsLine helper, and the
scroll-to-bottom tracker out of the task detail route. Brings the route
file from 503 LoC to 158 (target <250); each new component stays well
below 400.

The route still owns cross-section state — commentEffort, replyTarget,
the two comment-form refs, and the scrollParent useLayoutEffect — and
threads them through as props. scrollParent stays resolved at the route
because both CommentsSection (Virtuoso customScrollParent) and the new
useScrollToBottom hook consume it.
